Minor Scrypt tidy-ups. (#1055)
Update native method signatures to match ScryptKeySpec types, factor
constants out of tests and test all aliases at least once.

No functional change in this CL, more tests and implementation to
come in a future one.

+++ common/src/main/java/org/conscrypt/NativeCrypto.java
@@ -1446,7 +1446,7 @@ public final class NativeCrypto {
     /**
      * Generates a key from a password and salt using Scrypt.
      */
-    static native byte[] Scrypt_generate_key(byte[] password, byte[] salt, long n, long r, long p, int key_len);
+    static native byte[] Scrypt_generate_key(byte[] password, byte[] salt, int n, int r, int p, int key_len);
 
     /**
      * Return {@code true} if BoringSSL has been built in FIPS mode.
